Prompt for chromedriver install on Darwin (macOS)
When developing on non-Linux systems, the `make` target to set up all
development will also try to install a Linux-compiled version of
chromedriver.

On Darwin/macOS-based systems, the resulting file is not executable, and
running the test suite (specifically the acceptance tests) fails.

This introduces a condition to skip the installation of the binary on
Darwin-based systems, and prompts the user to install it via a commonly
used package management system.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>