feat(metrics) Add metrics for plugin API calls (#480)
Record metrics when we send requests to 3rd parties. Track the host and
status of the responses so we can better understand which plugins are
having problems and how severe those problems are.

Host names for plugins is a potentially unbounded set. Instead use the
plugin name as a tag. For plugins outside of this repository the name
will be `undefined` but that's ok for now, and can be improved in the
future if required.

Leveraging the base class lets us use less duplicate code and get
metrics/logging for free.

Refs SEN-630